Emily/secure apis

+++ controllers/profileController.js
@@ -1,9 +1,9 @@
 import Profile from '../models/profile.js';
 
-// GET
+// GET (only helpers)
 const getAllProfiles = async (req, res, next) => {
   try {
-    const allProfiles = await Profile.find();
+    const allProfiles = await Profile.find({ isHelper: true });
     return res.status(200).json({ status: 'success', body: allProfiles });
   } catch (err) {
     next(err);
@@ -14,7 +14,7 @@ const getAllProfiles = async (req, res, next) => {
 const searchProfile = async (req, res, next) => {
   try {
     console.log('req.params', req.params);
-    const allProfiles = await Profile.find();
+    const allProfiles = await Profile.find({ isHelper: true });
     console.log('allProfiles', allProfiles);
     const profileById = allProfiles.find((profile) => profile.id === req.params.searchTerm);
 
@@ -43,17 +43,23 @@ const searchProfile = async (req, res, next) => {
 };
 
 // PUT (no POST since that's register)
+// This will update ANY profile (user, helper, admin etc)
 const updateProfile = async (req, res, next) => {
   try {
-    // console.log('req.params', req.params);
-    // console.log('req.body', req.body);
-    const updatedProfile = await Profile.findById(req.params.id);
-    console.log({ updatedProfile });
-    if (!updatedProfile) {
+    console.log('req.currentUser._id', req.currentUser._id);
+    const profile = await Profile.findById(req.params.id);
+    console.log('profile._id', profile._id);
+    console.log({ profile });
+    if (!profile) {
       return res.status(400).json({ message: 'Profile not found' });
+    } else if (!req.currentUser._id.equals(profile._id)) {
+      return res.status(401).json({
+        message:
+          'Unauthorised action. You must be the owner of this user profile to update this profile.'
+      });
     } else {
-      updatedProfile.set(req.body);
-      const savedProfile = await updatedProfile.save();
+      profile.set(req.body);
+      const savedProfile = await profile.save();
       return res.status(200).json({ status: 'Successfully updated profile', body: savedProfile });
     }
   } catch (err) {
@@ -63,19 +69,25 @@ const updateProfile = async (req, res, next) => {
 
 // DEL (everyone can delete their own? or just admin? probably just admin)
 const deleteProfile = async (req, res, next) => {
-  try {
-    console.log('req.params', req.params);
-    const profile = await Profile.findById(req.params.id);
-    console.log('profile', profile);
-    if (!profile) {
-      return res.status(400).json({ message: 'Profile not found' });
-    } else {
-      await Profile.findByIdAndRemove(req.params.id);
-      return res.status(200).json({ message: 'Successfully deleted profile', body: profile });
+  if (req.currentUser.isAdmin) {
+    try {
+      console.log('req.params', req.params);
+      const profile = await Profile.findById(req.params.id);
+      console.log('profile', profile);
+      if (!profile) {
+        return res.status(400).json({ message: 'Profile not found' });
+      } else {
+        await Profile.findByIdAndRemove(req.params.id);
+        return res.status(200).json({ message: 'Successfully deleted profile', body: profile });
+      }
+    } catch (err) {
+      console.error(err);
+      next(err);
     }
-  } catch (err) {
-    console.error(err);
-    next(err);
+  } else {
+    return res.status(401).send({
+      message: 'Unauthorised: you must be an admin to delete a user profile.'
+    });
   }
 };
